Add ql.archtype, ql.ostype, ql.archendian etc
Move qling/loder/utils.py to qiling/utils.py
